finished updating fork handler
