+++ functions/fork/main.py
@@ -9,7 +9,7 @@ from __future__ import print_function
 
 import logging
 
-from handler import ForkHandler
+from libraries.lambda_handlers.fork_handler import ForkHandler
 from libraries.tools.file_utils import wipe_temp
 
 logger = logging.getLogger()
@@ -18,12 +18,6 @@ logger.setLevel(logging.INFO)
 
 def handle(event, context):
     wipe_temp(ignore_errors=True)
-    try:
-        handler = ForkHandler(event, logger)
-        handler.run()
-    except Exception as e:
-        raise Exception('Bad Request: {0}'.format(e))
 
-    return {
-        "success": True
-    }
+    handler = ForkHandler(event, context, logger=logger)
+    return handler.run()
